+++ snapatac2-python/snapatac2/datasets.py
@@ -6,6 +6,16 @@ import pooch
 
 from snapatac2._snapatac2 import read_motifs, PyDNAMotif
 
+def pbmc500() -> Path:
+    """500 PBMCs from 10x Genomics.
+
+    Returns
+    -------
+    Path
+        Path to the fragment file.
+    """
+    return Path(_datasets.fetch("atac_pbmc_500.tsv.gz"))
+
 def pbmc5k(type: Literal["fragment, h5ad, gene, annotated_h5ad"] = "fragment") -> Path:
     """5k PBMCs from 10x Genomics.
 
@@ -90,6 +100,7 @@ _datasets = pooch.create(
     base_url="http://renlab.sdsc.edu/kai/public_datasets/",
     # The registry specifies the files that can be fetched
     registry={
+        "atac_pbmc_500.tsv.gz": "sha256:196c5d7ee0169957417e9f4d5502abf1667ef99453328f8d290d4a7f3b205c6c",
         "atac_pbmc_5k.tsv.gz": "sha256:5fe44c0f8f76ce1534c1ae418cf0707ca5ef712004eee77c3d98d2d4b35ceaec",
         "atac_pbmc_5k.h5ad": "sha256:dcaca8ca4ac28674ec2172b4a975f75fba2ede1fc86571f7c452ba00f5cd4b94",
         "atac_pbmc_5k_annotated.h5ad": "sha256:3d5f147ce13a01cd2bdc3d9d2e8cf7897ee98e44255ff12f868517dd78427a87",
@@ -113,10 +124,13 @@ _datasets = pooch.create(
         "gencode_vM30_GRCm39.fa.gz": "sha256:3b923c06a0d291fe646af6bf7beaed7492bf0f6dd5309d4f5904623cab41b0aa",
     },
     urls={
+        "atac_pbmc_500.tsv.gz": "https://cf.10xgenomics.com/samples/cell-atac/2.0.0/atac_pbmc_500_nextgem/atac_pbmc_500_nextgem_fragments.tsv.gz",
+
         "atac_pbmc_5k.tsv.gz": "http://renlab.sdsc.edu/kai/public_datasets/single_cell_atac/atac_pbmc_5k_nextgem_fragments.tsv.gz",
         "atac_pbmc_5k.h5ad": "http://renlab.sdsc.edu/kai/public_datasets/single_cell_atac/atac_pbmc_5k.h5ad",
         "atac_pbmc_5k_gene.h5ad": "http://renlab.sdsc.edu/kai/public_datasets/single_cell_atac/atac_pbmc_5k_gene.h5ad",
         "atac_pbmc_5k_annotated.h5ad": "http://renlab.sdsc.edu/kai/public_datasets/single_cell_atac/atac_pbmc_5k_annotated.h5ad",
+
 
         "10x-Multiome-Pbmc10k-ATAC.h5ad": "http://renlab.sdsc.edu/kai/public_datasets/single_cell_multiome/10x-Multiome-Pbmc10k-ATAC.h5ad",
         "10x-Multiome-Pbmc10k-RNA.h5ad": "http://renlab.sdsc.edu/kai/public_datasets/single_cell_multiome/10x-Multiome-Pbmc10k-RNA.h5ad",